requirements.txt
sacred==0.7.5
sk-video
torch==1.1.0
torchvision
tensorboardX
dask[complete]
termcolor
argparse
tqdm
lmdb
Pillow
click
joblib
gym
numpy
matplotlib
filelock
pudb
pymongo
